+++ src/main/java/uk/ac/standrews/cs/population_linkage/groundTruth/AllPairsBirthMotherUmeaIdentityLinkage.java
@@ -25,7 +25,7 @@ public class AllPairsBirthMotherUmeaIdentityLinkage extends AllPairsSameSourceLi
 
     private static final int NUMBER_OF_RUNS = 1;
 
-    public AllPairsBirthMotherUmeaIdentityLinkage(Path store_path, String repo_name, String linkage_results_filename, final String distance_results_filename, int number_of_records_to_be_checked) throws IOException {
+    public AllPairsBirthMotherUmeaIdentityLinkage(Path store_path, String repo_name, String linkage_results_filename, final String distance_results_filename, long number_of_records_to_be_checked) throws IOException {
         super(store_path,repo_name,linkage_results_filename, distance_results_filename,number_of_records_to_be_checked,NUMBER_OF_RUNS);
     }
