+++ src/main/java/uk/ac/standrews/cs/population_linkage/groundTruth/AllPairsSameSourceLinkageAnalysis.java
@@ -47,10 +47,10 @@ public abstract class AllPairsSameSourceLinkageAnalysis extends ThresholdAnalysi
     private final Map<String, Integer> run_numbers_for_metrics;
 
     private List<LXP> source_records;
-    private int number_of_records;
-    private int records_processed = 0;
+    private long number_of_records;
+    private long records_processed = 0;
 
-    protected AllPairsSameSourceLinkageAnalysis(final Path store_path, final String repo_name, final String linkage_results_filename, final String distance_results_filename, int number_of_records_to_be_checked, int number_of_runs ) throws IOException {
+    protected AllPairsSameSourceLinkageAnalysis(final Path store_path, final String repo_name, final String linkage_results_filename, final String distance_results_filename, long number_of_records_to_be_checked, int number_of_runs ) throws IOException {
 
         super(number_of_records_to_be_checked, number_of_runs );
 
@@ -123,7 +123,7 @@ public abstract class AllPairsSameSourceLinkageAnalysis extends ThresholdAnalysi
         printHeaders();
         printMetaData();
 
-        final int number_of_blocks_to_be_checked = number_of_records / BLOCK_SIZE;
+        final long number_of_blocks_to_be_checked = number_of_records / BLOCK_SIZE;
 
         for (int block_index = 0; block_index < number_of_blocks_to_be_checked; block_index++) {
